requirements.txt
dm-control == 0.0.364896371
dm-robotics-moma == 0.0.4
dm-robotics-manipulation == 0.0.4
tensorflow == 2.7.0.rc0
dm-reverb-nightly == 0.5.0.dev20211104
opencv-python >= 3.4.0
typing-extensions >= 3.7.4
